Split views into separate files with tests
 * This is following the pattern agreed for LMS
 * Things in the view directory with _view(s) are views
 * Other things are just regular code
 * Split up the test files
 * Re-used the cache assertion code
 * Added a lint warning to the skips for tests
   * This prevents an incorrect warning on matchers usage
