+++ tests/.pylintrc
@@ -19,6 +19,7 @@ disable=bad-continuation,
         missing-type-doc,
         missing-return-type-doc,
         no-self-use,
+        no-value-for-parameter,
         redefined-outer-name,
         too-few-public-methods,
         unused-argument,
